dzil: add PkgVersion, Git::NextVersion and Git::Tag plugins
Not having a $VERSION in the package definition was causing it to be indexed
in 02packages w/o a version, causing cpanm to believe that no upgrade is
necessary even when an updated dist is available and indexed.

These plugins just add some nice and easy versioning while also allowing for
it to be overridden easily (e.g. "V=1.000 dzil release ...").