Drop SPF RR from generated output.
RFC 7208 section 3.1 states that the SPF RR type should no longer be
used and that TXT should be the only RR type used for SPF.
Modify code to only output TXT RR types.
We are still reading and parsing the SPF RR type, but the output
will be valid as per RFC 7208.